Revert "Add High-level Op Support (#5)"
This reverts commit 67960ee474c8797ca9111527f7b9eb6693377ad6.
