+++ include/tvm/relax/op_attr_types.h
@@ -109,34 +109,6 @@ struct AssertOpAttrs : public tvm::AttrsNode<AssertOpAttrs> {
   }
 };
 
-/*! \brief Attributes used in MaxPool2d operator */
-struct MaxPool2dAttrs : public tvm::AttrsNode<MaxPool2dAttrs> {
-  Array<PrimExpr> kernel_size;
-  Array<PrimExpr> stride;
-  Array<PrimExpr> padding;
-  Array<PrimExpr> dilation;
-  TVM_DECLARE_ATTRS(MaxPool2dAttrs, "relax.attrs.MaxPool2dAttrs") {
-    TVM_ATTR_FIELD(kernel_size).describe("The size of the window to take a max over.");
-    TVM_ATTR_FIELD(stride).describe("The stride of the window.");
-    TVM_ATTR_FIELD(padding).describe("The padding on the input.");
-    TVM_ATTR_FIELD(dilation).describe("The stride of elements in the window.");
-  }
-};  // struct MaxPool2dAttrs
-
-/*! \brief Attributes used in Conv2d operator */
-struct Conv2dAttrs : public tvm::AttrsNode<Conv2dAttrs> {
-  Array<PrimExpr> kernel_size;
-  Array<PrimExpr> stride;
-  Array<PrimExpr> padding;
-  Array<PrimExpr> dilation;
-  TVM_DECLARE_ATTRS(Conv2dAttrs, "relax.attrs.Conv2dAttrs") {
-    TVM_ATTR_FIELD(kernel_size).describe("The size of the convolving kernel.");
-    TVM_ATTR_FIELD(stride).describe("The stride of the convolution.");
-    TVM_ATTR_FIELD(padding).describe("The padding on the input.");
-    TVM_ATTR_FIELD(dilation).describe("The spacing between kernel elements.");
-  }
-};  // struct Conv2dAttrs
-
 }  // namespace relax
 }  // namespace tvm
 #endif  // TVM_RELAX_OP_ATTR_TYPES_H_
