--- src/relax/op/nn/nn.cc
@@ -1,56 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-#include "nn.h"
-
-namespace tvm {
-namespace relax {
-
-RELAY_REGISTER_OP("relax.nn.dense")
-    .set_num_inputs(2)
-    .add_argument("e1", "Expr", "The input expression")
-    .add_argument("e2", "Expr", "The input expression")
-    .set_attr<FInferShape>("FInferShape", InferShapeDense)
-    .set_attr<FInferType>("FInferType", InferTypeDense);
-
-Expr MakeDense(Expr expr1, Expr expr2) {
-  static const Op& op = Op::Get("relax.nn.dense");
-  return Call(op, {expr1, expr2}, {}, {});
-}
-
-TVM_REGISTER_GLOBAL("relax.op.nn.dense").set_body_typed(MakeDense);
-
-RELAX_REGISTER_UNARY_OP("nn.softmax");
-
-RELAX_REGISTER_UNARY_OP("nn.relu");
-
-RELAY_REGISTER_OP("relax.nn.flatten")
-    .set_num_inputs(1)
-    .add_argument("data", "Tensor", "The input tensor")
-    .set_attr<FInferShape>("FInferShape", InferShapeFlatten)
-    .set_attr<FInferType>("FInferType", InferTypeFlatten);
-
-Expr MakeFlatten(Expr data) {
-  static const Op& op = Op::Get("relax.nn.flatten");
-  return Call(op, {data}, {}, {});
-}
-TVM_REGISTER_GLOBAL("relax.op.nn.flatten").set_body_typed(MakeFlatten);
-
-}  // namespace relax
-}  // namespace tvm
