+++ src/relax/op/tensor/unary.cc
@@ -27,66 +27,8 @@
 namespace tvm {
 namespace relax {
 
-Optional<Expr> InferShapeUnique(const Call& call, DiagnosticContext diag_ctx) {
-  if (call->args.size() != 1) {
-    diag_ctx.EmitFatal(Diagnostic::Error(call->span) << "Unique op should have 1 argument");
-  }
-  auto unique_attrs = call->attrs.as<UniqueAttrs>();
-  // Only default values of these attributes are supported right now.
-  if (unique_attrs->return_counts || unique_attrs->return_inverse || unique_attrs->dim != -1)
-    diag_ctx.EmitFatal(Diagnostic::Error(call->span)
-                       << "support for return_inverse, return_counts, and dim is not implemented");
-  return relax::RuntimeDepShape(call->span);
-}
-
-Type InferTypeUnique(const Call& call, DiagnosticContext diag_ctx) {
-  if (call->args.size() != 1) {
-    diag_ctx.EmitFatal(Diagnostic::Error(call->span) << "Unique op should have 1 argument");
-  }
-  auto* input_ty = call->args[0]->checked_type().as<DynTensorTypeNode>();
-  if (!input_ty) {
-    diag_ctx.EmitFatal(Diagnostic::Error(call->span)
-                       << "Input should be DynTensor, but got "
-                       << call->args[0]->checked_type()->GetTypeKey());
-  }
-
-  // TODO(prakalp): Add support for return_inverse, return_counts and dim attributes. Only defaults
-  // are supported right now.
-  auto unique_attrs = call->attrs.as<UniqueAttrs>();
-  if (unique_attrs->return_counts || unique_attrs->return_inverse || unique_attrs->dim != -1)
-    diag_ctx.EmitFatal(Diagnostic::Error(call->span)
-                       << "support for return_inverse, return_counts, and dim is not implemented");
-  return DynTensorType(/*ndim=*/1, input_ty->dtype);
-}
-
-Optional<Expr> InferShapeUnaryBroadcast(const Call& call, DiagnosticContext diag_ctx) {
-  if (call->args.size() != 1) {
-    diag_ctx.EmitFatal(Diagnostic::Error(call->span) << "Unary op should have 1 argument");
-  }
-  Expr shape = call->args[0]->shape();
-  auto* s = shape.as<ShapeExprNode>();
-  if (s) {
-    return ShapeExpr(s->values);
-  } else {
-    return NullOpt;
-  }
-}
-
-Type InferTypeUnaryBroadcast(const Call& call, DiagnosticContext diag_ctx) {
-  if (call->args.size() != 1) {
-    diag_ctx.EmitFatal(Diagnostic::Error(call->span) << "Unary op should have 1 argument");
-  }
-  auto* input_ty = call->args[0]->checked_type().as<DynTensorTypeNode>();
-  if (!input_ty) {
-    diag_ctx.EmitFatal(Diagnostic::Error(call->span)
-                       << "Input should be DynTensor, but got "
-                       << call->args[0]->checked_type()->GetTypeKey());
-  }
-  return GetRef<DynTensorType>(input_ty);
-}
-
 TVM_REGISTER_NODE_TYPE(UniqueAttrs);
-TVM_REGISTER_NODE_TYPE(MaxPool2dAttrs);
+
 RELAY_REGISTER_OP("relax.unique")
     .describe(
         "This operation returns the unique elements and the new index of each item in a given "
